+++ src/main/java/roomescape/infrastructure/auth/jwt/JwtCookieResolver.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.infrastructure.auth.jwt;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
