+++ src/main/java/roomescape/infrastructure/auth/member/UserInfo.java
@@ -1,4 +1,4 @@
-package roomescape.auth.member;
+package roomescape.infrastructure.auth.member;
 
 import roomescape.domain.member.Role;
