+++ src/test/java/roomescape/RoomescapeApplicationTest.java
@@ -3,7 +3,7 @@ package roomescape;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
-import roomescape.auth.config.WebMvcConfiguration;
+import roomescape.infrastructure.auth.config.WebMvcConfiguration;
 
 @SpringBootTest
 @Import(WebMvcConfiguration.class)
