+++ src/test/java/roomescape/unit/service/reservationmember/ReservationMemberServiceTest.java
@@ -11,12 +11,12 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import roomescape.auth.JwtTokenProvider;
 import roomescape.domain.reservationmember.ReservationMember;
 import roomescape.dto.member.SignupRequestDto;
 import roomescape.dto.reservation.AddReservationDto;
 import roomescape.dto.reservation.AddReservationTimeDto;
 import roomescape.dto.reservation.AddThemeDto;
+import roomescape.infrastructure.auth.jwt.JwtTokenProvider;
 import roomescape.service.member.MemberService;
 import roomescape.service.reservation.ReservationService;
 import roomescape.service.reservation.ReservationTimeService;
